requirements.txt
asgiref==3.8.1
channels==4.1.0
Django==5.0.6
django_channels==0.7.0
django-cors-headers==4.3.1
djangorestframework==3.15.1
channels-redis==4.2.0
matplotlib==3.9.0
numpy==1.24.3
pandas==2.2.2
pygame==2.5.2
redis==5.0.4
torch==2.3.0+cu118
tqdm==4.66.4
